max dirs working

+++ messaging/msgq.cc
@@ -25,7 +25,7 @@
 
 #include "msgq.h"
 
-int MAX_MSG_DIRS = 5;
+int MAX_MSG_DIRS = 200;
 
 void sigusr2_handler(int signal) {
   assert(signal == SIGUSR2);
@@ -88,19 +88,14 @@ void msgq_wait_for_subscriber(msgq_queue_t *q){
 void clean_msg_dirs(std::string path){
   path.pop_back();
   std::filesystem::directory_iterator iter(path.substr(0, path.rfind("/")));
-  auto iter_b = std::filesystem::begin(iter);
-  auto iter_e = std::filesystem::end(iter);
-  if (std::distance(iter_b, iter_e) > MAX_MSG_DIRS) {
-    std::optional<std::filesystem::directory_entry> oldest;
-    for (std::filesystem::directory_entry entry : iter_b) {
-      if (!oldest.has_value()) oldest = entry;
-      else if (entry.last_write_time() < oldest->last_write_time()) {
-        oldest = entry;
-      }
-      std::cout<<oldest->path()<<std::endl;
-    }
-    std::filesystem::remove_all(oldest->path());
+  int count = 0;
+  std::optional<std::filesystem::directory_entry> oldest;
+  for (std::filesystem::directory_entry entry : iter) {
+    count += 1;
+    if (!oldest.has_value()) oldest = entry;
+    else if (entry.last_write_time() < oldest->last_write_time()) oldest = entry;
   }
+  if (count > MAX_MSG_DIRS) std::filesystem::remove_all(oldest->path());
 }
 
 int msgq_new_queue(msgq_queue_t * q, const char * path, size_t size){
@@ -110,7 +105,7 @@ int msgq_new_queue(msgq_queue_t * q, const char * path, size_t size){
   std::string prefix_str = "/dev/shm/"+std::string(std::getenv("OPENPILOIT_PREFIX"))+"/";
   const char * prefix = prefix_str.c_str();
   std::filesystem::create_directory(prefix);
-  //clean_msg_dirs(prefix_str);
+  clean_msg_dirs(prefix_str);
   char * full_path = new char[strlen(path) + strlen(prefix) + 1];
   strcpy(full_path, prefix);
   strcat(full_path, path);
